using float() after divide
In Python, if one "int" number is divided by another "int" number, the type of the result is "int", unless one's type is "float".
So I think it's better to replace "ratio = float(input_width / c)" with "ratio = float(input_width) / c".
Also, replace "ratio = float(32 / r)" with "ratio = 32.0 / r".

Therefore, when testing, the paper said the width are proportionally with heights, but at least 100 pixels. I"m trying to fix this, is the idea right?